+++ src/Foundry.cpp
@@ -12,6 +12,7 @@
 #include <algorithm>
 #include <time.h>
 #include "FoundrySequencer.hpp"
+#include "comp/PianoKey.hpp"
 
 
 struct Foundry : Module {	
@@ -27,7 +28,7 @@ struct Foundry : Module {
 		GATE_PARAM,
 		SLIDE_BTN_PARAM,
 		AUTOSTEP_PARAM,
-		ENUMS(KEY_PARAMS, 12),
+		ENUMS(KEY_PARAMS, 12),// no longer used
 		MODE_PARAM,
 		CLKRES_PARAM,
 		TRAN_ROT_PARAM,
@@ -136,7 +137,7 @@ struct Foundry : Module {
 	Trigger rightTrigger;
 	Trigger runningTrigger;
 	Trigger clockTriggers[SequencerKernel::MAX_STEPS];
-	Trigger keyTriggers[12];
+	dsp::BooleanTrigger keyTrigger;
 	Trigger octTriggers[7];
 	Trigger gate1Trigger;
 	Trigger tiedTrigger;
@@ -161,6 +162,7 @@ struct Foundry : Module {
 	Trigger allTrigger;
 	Trigger velEditTrigger;
 	Trigger writeModeTrigger;
+	PianoKeyInfo pkInfo;
 
 	
 	inline bool isEditingSequence(void) {return params[EDIT_PARAM].getValue() > 0.5f;}
@@ -200,19 +202,6 @@ struct Foundry : Module {
 			snprintf(strBuf, 32, "Octave %i", i + 1);
 			configParam(OCTAVE_PARAM + i, 0.0f, 1.0f, 0.0f, strBuf);
 		}
-		configParam(KEY_PARAMS + 1, 0.0, 1.0, 0.0, "C# key");
-		configParam(KEY_PARAMS + 3, 0.0, 1.0, 0.0, "D# key");
-		configParam(KEY_PARAMS + 6, 0.0, 1.0, 0.0, "F# key");
-		configParam(KEY_PARAMS + 8, 0.0, 1.0, 0.0, "G# key");
-		configParam(KEY_PARAMS + 10, 0.0, 1.0, 0.0, "A# key");
-
-		configParam(KEY_PARAMS + 0, 0.0, 1.0, 0.0, "C key");
-		configParam(KEY_PARAMS + 2, 0.0, 1.0, 0.0, "D key");
-		configParam(KEY_PARAMS + 4, 0.0, 1.0, 0.0, "E key");
-		configParam(KEY_PARAMS + 5, 0.0, 1.0, 0.0, "F key");
-		configParam(KEY_PARAMS + 7, 0.0, 1.0, 0.0, "G key");
-		configParam(KEY_PARAMS + 9, 0.0, 1.0, 0.0, "A key");
-		configParam(KEY_PARAMS + 11, 0.0, 1.0, 0.0, "B key");
 
 		configParam(VEL_KNOB_PARAM, -INFINITY, INFINITY, 0.0f, "CV2/p/r knob");	
 		configParam(VEL_EDIT_PARAM, 0.0f, 1.0f, 0.0f, "CV2/p/r select");
@@ -902,20 +891,17 @@ struct Foundry : Module {
 			}
 			
 			// Keyboard buttons
-			for (int keyn = 0; keyn < 12; keyn++) {
-				if (keyTriggers[keyn].process(params[KEY_PARAMS + keyn].getValue())) {
-					if (editingSequence) {
-						displayState = DISP_NORMAL;
-						bool autostepClick = paramQuantities[KEY_PARAMS + keyn]->getMaxValue() > 1.5f;// if double-click
-						if (isEditingGates()) {
-							if (!seq.setGateType(keyn, multiSteps ? cpSeqLength : 1, sampleRate, autostepClick, multiTracks))
-								displayState = DISP_PPQN;
-						}
-						else {
-							if (seq.applyNewKey(keyn, multiSteps ? cpSeqLength : 1, sampleRate, autostepClick, multiTracks))
-								tiedWarning = (long) (warningTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
-						}							
+			if (keyTrigger.process(pkInfo.gate)) {
+				if (editingSequence) {
+					displayState = DISP_NORMAL;
+					if (isEditingGates()) {
+						if (!seq.setGateType(pkInfo.key, multiSteps ? cpSeqLength : 1, sampleRate, pkInfo.isRightClick, multiTracks))
+							displayState = DISP_PPQN;
 					}
+					else {
+						if (seq.applyNewKey(pkInfo.key, multiSteps ? cpSeqLength : 1, sampleRate, pkInfo.isRightClick, multiTracks))
+							tiedWarning = (long) (warningTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
+					}							
 				}
 			}
 			
@@ -1931,30 +1917,30 @@ struct FoundryWidget : ModuleWidget {
 		static const int offsetKeyLEDx = 6;
 		static const int offsetKeyLEDy = 16;
 		// Black keys and lights
-		addParam(createParam<InvisibleKeySmall>(			Vec(65+keyNudgeX, KeyBlackY), module, Foundry::KEY_PARAMS + 1));
+		addChild(createPianoKey<PianoKeySmall>(Vec(65+keyNudgeX, KeyBlackY), 1, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(65+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 1 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(93+keyNudgeX, KeyBlackY), module, Foundry::KEY_PARAMS + 3));
+		addChild(createPianoKey<PianoKeySmall>(Vec(93+keyNudgeX, KeyBlackY), 3, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(93+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 3 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(150+keyNudgeX, KeyBlackY), module, Foundry::KEY_PARAMS + 6));
+		addChild(createPianoKey<PianoKeySmall>(Vec(150+keyNudgeX, KeyBlackY), 6, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(150+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 6 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(178+keyNudgeX, KeyBlackY), module, Foundry::KEY_PARAMS + 8));
+		addChild(createPianoKey<PianoKeySmall>(Vec(178+keyNudgeX, KeyBlackY), 8, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(178+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 8 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(206+keyNudgeX, KeyBlackY), module, Foundry::KEY_PARAMS + 10));
+		addChild(createPianoKey<PianoKeySmall>(Vec(206+keyNudgeX, KeyBlackY), 10, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(206+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 10 * 2));
 		// White keys and lights
-		addParam(createParam<InvisibleKeySmall>(			Vec(51+keyNudgeX, KeyWhiteY), module, Foundry::KEY_PARAMS + 0));
+		addChild(createPianoKey<PianoKeySmall>(Vec(51+keyNudgeX, KeyWhiteY), 0, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(51+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 0 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(79+keyNudgeX, KeyWhiteY), module, Foundry::KEY_PARAMS + 2));
+		addChild(createPianoKey<PianoKeySmall>(Vec(79+keyNudgeX, KeyWhiteY), 2, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(79+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 2 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(107+keyNudgeX, KeyWhiteY), module, Foundry::KEY_PARAMS + 4));
+		addChild(createPianoKey<PianoKeySmall>(Vec(107+keyNudgeX, KeyWhiteY), 4, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(107+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 4 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(136+keyNudgeX, KeyWhiteY), module, Foundry::KEY_PARAMS + 5));
+		addChild(createPianoKey<PianoKeySmall>(Vec(136+keyNudgeX, KeyWhiteY), 5, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(136+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 5 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(164+keyNudgeX, KeyWhiteY), module, Foundry::KEY_PARAMS + 7));
+		addChild(createPianoKey<PianoKeySmall>(Vec(164+keyNudgeX, KeyWhiteY), 7, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(164+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 7 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(192+keyNudgeX, KeyWhiteY), module, Foundry::KEY_PARAMS + 9));
+		addChild(createPianoKey<PianoKeySmall>(Vec(192+keyNudgeX, KeyWhiteY), 9, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(192+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 9 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(220+keyNudgeX, KeyWhiteY), module, Foundry::KEY_PARAMS + 11));
+		addChild(createPianoKey<PianoKeySmall>(Vec(220+keyNudgeX, KeyWhiteY), 11, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(220+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, Foundry::KEY_LIGHTS + 11 * 2));
 
 
@@ -2138,12 +2124,3 @@ struct FoundryWidget : ModuleWidget {
 };
 
 Model *modelFoundry = createModel<Foundry, FoundryWidget>("Foundry");
-
-/*CHANGE LOG
-
-1.0.0:
-expansion panel replaced by a separate expander module
-right-click keys to autostep replaced by double click
-add menu option to stop at end of song
-
-*/