+++ src/GateSeq64.cpp
@@ -1594,13 +1594,3 @@ struct GateSeq64Widget : ModuleWidget {
 };
 
 Model *modelGateSeq64 = createModel<GateSeq64, GateSeq64Widget>("Gate-Seq-64");
-
-/*CHANGE LOG
-
-1.0.0:
-removed right-click of step buttons to clear gates
-expansion panel replaced by a separate expander module
-add menu option to stop at end of song
-implement mouse painting and change click strategy in the steps 
-
-*/