+++ src/PhraseSeq16.cpp
@@ -13,6 +13,7 @@
 
 
 #include "PhraseSeqUtil.hpp"
+#include "comp/PianoKey.hpp"
 
 
 struct PhraseSeq16 : Module {
@@ -36,7 +37,7 @@ struct PhraseSeq16 : Module {
 		ROTATER_PARAM,// no longer used
 		PASTESYNC_PARAM,// no longer used
 		AUTOSTEP_PARAM,
-		ENUMS(KEY_PARAMS, 12),
+		ENUMS(KEY_PARAMS, 12),// no longer used
 		TRANSPOSEU_PARAM,// no longer used
 		TRANSPOSED_PARAM,// no longer used
 		RUNMODE_PARAM,
@@ -165,7 +166,7 @@ struct PhraseSeq16 : Module {
 	Trigger gate1ProbTrigger;
 	Trigger gate2Trigger;
 	Trigger slideTrigger;
-	Trigger keyTriggers[12];
+	dsp::BooleanTrigger keyTrigger;
 	Trigger writeTrigger;
 	Trigger attachedTrigger;
 	Trigger copyTrigger;
@@ -179,6 +180,7 @@ struct PhraseSeq16 : Module {
 	Trigger keyGateTrigger;
 	Trigger seqCVTrigger;
 	HoldDetect modeHoldDetect;
+	PianoKeyInfo pkInfo;
 
 	
 	inline bool isEditingSequence(void) {return params[EDIT_PARAM].getValue() > 0.5f;}
@@ -205,19 +207,6 @@ struct PhraseSeq16 : Module {
 			snprintf(strBuf, 32, "Octave %i", i + 1);
 			configParam(OCTAVE_PARAM + i, 0.0f, 1.0f, 0.0f, strBuf);
 		}
-		configParam(KEY_PARAMS + 1, 0.0, 1.0, 0.0, "C# key");
-		configParam(KEY_PARAMS + 3, 0.0, 1.0, 0.0, "D# key");
-		configParam(KEY_PARAMS + 6, 0.0, 1.0, 0.0, "F# key");
-		configParam(KEY_PARAMS + 8, 0.0, 1.0, 0.0, "G# key");
-		configParam(KEY_PARAMS + 10, 0.0, 1.0, 0.0, "A# key");
-
-		configParam(KEY_PARAMS + 0, 0.0, 1.0, 0.0, "C key");
-		configParam(KEY_PARAMS + 2, 0.0, 1.0, 0.0, "D key");
-		configParam(KEY_PARAMS + 4, 0.0, 1.0, 0.0, "E key");
-		configParam(KEY_PARAMS + 5, 0.0, 1.0, 0.0, "F key");
-		configParam(KEY_PARAMS + 7, 0.0, 1.0, 0.0, "G key");
-		configParam(KEY_PARAMS + 9, 0.0, 1.0, 0.0, "A key");
-		configParam(KEY_PARAMS + 11, 0.0, 1.0, 0.0, "B key");
 		
 		configParam(EDIT_PARAM, 0.0f, 1.0f, 1.0f, "Seq/song mode");
 		configParam(RUNMODE_PARAM, 0.0f, 1.0f, 0.0f, "Length / run mode");
@@ -1077,47 +1066,45 @@ struct PhraseSeq16 : Module {
 			}
 			
 			// Keyboard buttons
-			for (int i = 0; i < 12; i++) {
-				if (keyTriggers[i].process(params[KEY_PARAMS + i].getValue())) {
-					if (editingSequence) {
-						displayState = DISP_NORMAL;
-						if (editingGateLength != 0l) {
-							int newMode = keyIndexToGateMode(i, pulsesPerStep);
-							if (newMode != -1) {
-								editingPpqn = 0l;
-								attributes[seqIndexEdit][stepIndexEdit].setGateMode(newMode, editingGateLength > 0l);
-								if (paramQuantities[KEY_PARAMS + i]->getMaxValue() > 1.5f) {// if double-click
-									stepIndexEdit = moveIndex(stepIndexEdit, stepIndexEdit + 1, 16);
-									editingType = (unsigned long) (gateTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
-									editingGateKeyLight = i;
-									if ((APP->window->getMods() & RACK_MOD_MASK) == RACK_MOD_CTRL)
-										attributes[seqIndexEdit][stepIndexEdit].setGateMode(newMode, editingGateLength > 0l);
-								}
-							}
-							else
-								editingPpqn = (long) (editGateLengthTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
-						}
-						else if (attributes[seqIndexEdit][stepIndexEdit].getTied()) {
-							if (paramQuantities[KEY_PARAMS + i]->getMaxValue() > 1.5f)// if double-click
+			if (keyTrigger.process(pkInfo.gate)) {
+				if (editingSequence) {
+					displayState = DISP_NORMAL;
+					if (editingGateLength != 0l) {
+						int newMode = keyIndexToGateMode(pkInfo.key, pulsesPerStep);
+						if (newMode != -1) {
+							editingPpqn = 0l;
+							attributes[seqIndexEdit][stepIndexEdit].setGateMode(newMode, editingGateLength > 0l);
+							if (pkInfo.isRightClick) {
 								stepIndexEdit = moveIndex(stepIndexEdit, stepIndexEdit + 1, 16);
-							else
-								tiedWarning = (long) (warningTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
-						}
-						else {	
-							float newCV = std::floor(cv[seqIndexEdit][stepIndexEdit]) + ((float) i) / 12.0f;
-							cv[seqIndexEdit][stepIndexEdit] = newCV;
-							propagateCVtoTied(seqIndexEdit, stepIndexEdit);
-							editingGate = (unsigned long) (gateTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
-							editingGateCV = cv[seqIndexEdit][stepIndexEdit];
-							editingGateKeyLight = -1;
-							if (paramQuantities[KEY_PARAMS + i]->getMaxValue() > 1.5f) {// if double-click
-								stepIndexEdit = moveIndex(stepIndexEdit, stepIndexEdit + 1, 16);
-								editingGateKeyLight = i;
+								editingType = (unsigned long) (gateTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
+								editingGateKeyLight = pkInfo.key;
 								if ((APP->window->getMods() & RACK_MOD_MASK) == RACK_MOD_CTRL)
-									cv[seqIndexEdit][stepIndexEdit] = newCV;
+									attributes[seqIndexEdit][stepIndexEdit].setGateMode(newMode, editingGateLength > 0l);
 							}
-						}						
+						}
+						else
+							editingPpqn = (long) (editGateLengthTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
 					}
+					else if (attributes[seqIndexEdit][stepIndexEdit].getTied()) {
+						if (pkInfo.isRightClick)
+							stepIndexEdit = moveIndex(stepIndexEdit, stepIndexEdit + 1, 16);
+						else
+							tiedWarning = (long) (warningTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
+					}
+					else {	
+						float newCV = std::floor(cv[seqIndexEdit][stepIndexEdit]) + ((float) pkInfo.key) / 12.0f;
+						cv[seqIndexEdit][stepIndexEdit] = newCV;
+						propagateCVtoTied(seqIndexEdit, stepIndexEdit);
+						editingGate = (unsigned long) (gateTime * sampleRate / RefreshCounter::displayRefreshStepSkips);
+						editingGateCV = cv[seqIndexEdit][stepIndexEdit];
+						editingGateKeyLight = -1;
+						if (pkInfo.isRightClick) {
+							stepIndexEdit = moveIndex(stepIndexEdit, stepIndexEdit + 1, 16);
+							editingGateKeyLight = pkInfo.key;
+							if ((APP->window->getMods() & RACK_MOD_MASK) == RACK_MOD_CTRL)
+								cv[seqIndexEdit][stepIndexEdit] = newCV;
+						}
+					}						
 				}
 			}
 					
@@ -1905,30 +1892,30 @@ struct PhraseSeq16Widget : ModuleWidget {
 		static const int offsetKeyLEDx = 6;
 		static const int offsetKeyLEDy = 16;
 		// Black keys and lights
-		addParam(createParam<InvisibleKeySmall>(			Vec(65+keyNudgeX, KeyBlackY), module, PhraseSeq16::KEY_PARAMS + 1));
+		addChild(createPianoKey<PianoKeySmall>(Vec(65+keyNudgeX, KeyBlackY), 1, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(65+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 1 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(93+keyNudgeX, KeyBlackY), module, PhraseSeq16::KEY_PARAMS + 3));
+		addChild(createPianoKey<PianoKeySmall>(Vec(93+keyNudgeX, KeyBlackY), 3, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(93+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 3 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(150+keyNudgeX, KeyBlackY), module, PhraseSeq16::KEY_PARAMS + 6));
+		addChild(createPianoKey<PianoKeySmall>(Vec(150+keyNudgeX, KeyBlackY), 6, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(150+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 6 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(178+keyNudgeX, KeyBlackY), module, PhraseSeq16::KEY_PARAMS + 8));
+		addChild(createPianoKey<PianoKeySmall>(Vec(178+keyNudgeX, KeyBlackY), 8, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(178+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 8 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(206+keyNudgeX, KeyBlackY), module, PhraseSeq16::KEY_PARAMS + 10));
+		addChild(createPianoKey<PianoKeySmall>(Vec(206+keyNudgeX, KeyBlackY), 10, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(206+keyNudgeX+offsetKeyLEDx, KeyBlackY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 10 * 2));
 		// White keys and lights
-		addParam(createParam<InvisibleKeySmall>(			Vec(51+keyNudgeX, KeyWhiteY), module, PhraseSeq16::KEY_PARAMS + 0));
+		addChild(createPianoKey<PianoKeySmall>(Vec(51+keyNudgeX, KeyWhiteY), 0, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(51+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 0 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(79+keyNudgeX, KeyWhiteY), module, PhraseSeq16::KEY_PARAMS + 2));
+		addChild(createPianoKey<PianoKeySmall>(Vec(79+keyNudgeX, KeyWhiteY), 2, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(79+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 2 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(107+keyNudgeX, KeyWhiteY), module, PhraseSeq16::KEY_PARAMS + 4));
+		addChild(createPianoKey<PianoKeySmall>(Vec(107+keyNudgeX, KeyWhiteY), 4, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(107+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 4 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(136+keyNudgeX, KeyWhiteY), module, PhraseSeq16::KEY_PARAMS + 5));
+		addChild(createPianoKey<PianoKeySmall>(Vec(136+keyNudgeX, KeyWhiteY), 5, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(136+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 5 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(164+keyNudgeX, KeyWhiteY), module, PhraseSeq16::KEY_PARAMS + 7));
+		addChild(createPianoKey<PianoKeySmall>(Vec(164+keyNudgeX, KeyWhiteY), 7, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(164+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 7 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(192+keyNudgeX, KeyWhiteY), module, PhraseSeq16::KEY_PARAMS + 9));
+		addChild(createPianoKey<PianoKeySmall>(Vec(192+keyNudgeX, KeyWhiteY), 9, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(192+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 9 * 2));
-		addParam(createParam<InvisibleKeySmall>(			Vec(220+keyNudgeX, KeyWhiteY), module, PhraseSeq16::KEY_PARAMS + 11));
+		addChild(createPianoKey<PianoKeySmall>(Vec(220+keyNudgeX, KeyWhiteY), 11, module ? &module->pkInfo : NULL));
 		addChild(createLight<MediumLight<GreenRedLight>>(Vec(220+keyNudgeX+offsetKeyLEDx, KeyWhiteY+offsetKeyLEDy), module, PhraseSeq16::KEY_LIGHTS + 11 * 2));
 		
 		
@@ -2054,12 +2041,3 @@ struct PhraseSeq16Widget : ModuleWidget {
 };
 
 Model *modelPhraseSeq16 = createModel<PhraseSeq16, PhraseSeq16Widget>("Phrase-Seq-16");
-
-/*CHANGE LOG
-
-1.0.0:
-expansion panel replaced by a separate expander module
-right-click keys to autostep replaced by double click
-add menu option to stop at end of song
-
-*/