+++ src/WriteSeq64.cpp
@@ -837,10 +837,3 @@ struct WriteSeq64Widget : ModuleWidget {
 };
 
 Model *modelWriteSeq64 = createModel<WriteSeq64, WriteSeq64Widget>("Write-Seq-64");
-
-/*CHANGE LOG
-
-1.0.0:
-removed right-click of gate button to clear gate
-
-*/