+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-### 1.0.2 (in development)
+### 1.1.0 (in development)
 
-- Added option in right-click menu to inverty velocity range
+- Added option in right-click menu to inverty velocity range in TwelveKey
+- Reinstated right-click of keys to perform autostep when entering notes in Foundry, PhraseSeq16/32 and SemiModularSynth (ctrl-right-click does the same but also copies the note over to new step)
 
 
 ### 1.0.1 (2019-06-23)